# .specify/specs/002-account-auth-onboarding/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: News Pulse — 계정·인증·온보딩·인가 API
  version: 1.0.0
  description: |
    spec 002 "계정·인증·온보딩·인가" API Contract.
    모든 시각은 UTC ISO-8601. 액세스 토큰 TTL 1h, 리프레시 토큰 TTL 30d.
    `/api/v1/admin/**` — ADMIN 역할 전용.
    `/api/v1/**` (나머지) — 인증 필수 (이메일 인증 완료 포함).
    공개 화이트리스트: 인증·비밀번호 재설정·헬스체크 엔드포인트. 이메일 인증 엔드포인트는 인증 필수(토큰 필요, emailVerified=false 허용).
servers:
  - url: http://localhost:8080
    description: 로컬 개발
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      description: RFC 7807 Problem Details 기반 에러 응답
      properties:
        code:
          type: string
          example: INVALID_PASSWORD
        message:
          type: string
          example: 비밀번호는 영문·숫자를 포함한 8자 이상이어야 합니다
        timestamp:
          type: string
          format: date-time
          example: '2026-06-11T10:00:00Z'
      required: [code, message, timestamp]
    TokenPair:
      type: object
      description: 인증 세션 토큰 쌍
      properties:
        accessToken:
          type: string
          description: JWT 액세스 토큰 (TTL 1h)
          example: eyJhbGciOiJIUzI1NiJ9...
        refreshToken:
          type: string
          description: 리프레시 토큰 (TTL 30d). Rotation 적용 — 사용 시 교체됨
          example: a1b2c3d4...
        expiresIn:
          type: integer
          description: 액세스 토큰 만료까지 남은 초
          example: 3600
      required: [accessToken, refreshToken, expiresIn]
    AccountSummary:
      type: object
      description: 현재 인증된 계정 요약 정보
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
          description: 소셜 전용 계정은 null 가능
        role:
          type: string
          enum: [USER, ADMIN]
        emailVerified:
          type: boolean
        onboardingCompleted:
          type: boolean
        signupType:
          type: string
          enum: [EMAIL, SOCIAL]
        createdAt:
          type: string
          format: date-time
        requiresReConsent:
          type: boolean
          description: 현재 활성 필수 약관 중 미동의 버전 존재 여부. true이면 클라이언트가 재동의 UI를 표시해야 함
      required: [id, role, emailVerified, onboardingCompleted, signupType, createdAt, requiresReConsent]

    # ── Signup / Login ───────────────────────────────────────────────────────
    SignupRequest:
      type: object
      description: 이메일 회원가입 요청
      properties:
        email:
          type: string
          format: email
          description: 가입 이메일 (대소문자 불구분 저장)
          example: user@example.com
        password:
          type: string
          description: 영문·숫자 각 1자 이상 포함, 8자 이상
          minLength: 8
          example: Password1
        consents:
          type: array
          description: 약관 동의 목록 (필수 약관 미동의 시 422)
          items:
            $ref: '#/components/schemas/ConsentInput'
        ageConfirmed:
          type: boolean
          description: 만 14세 이상 동의 여부. false이면 422
          example: true
      required: [email, password, consents, ageConfirmed]
    ConsentInput:
      type: object
      properties:
        termsVersionId:
          type: string
          format: uuid
          description: 동의할 약관 버전 ID
        agreed:
          type: boolean
      required: [termsVersionId, agreed]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: 현재 유효한 리프레시 토큰
      required: [refreshToken]
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: 로그아웃할 세션의 리프레시 토큰 (해당 기기 세션만 무효화)
      required: [refreshToken]

    # ── Password Reset ────────────────────────────────────────────────────────
...
    EmailVerificationRequestBody:
      type: object
      description: |
        이메일 인증 코드 발송 요청.
        가입 직후 자동 발송되며, 이 엔드포인트는 재발송 요청에 사용.
      properties:
        email:
          type: string
          format: email
          description: 인증할 이메일 (현재 계정 이메일과 일치해야 함)
      required: [email]
    EmailVerificationVerifyRequest:
      type: object
      properties:
        code:
          type: string
          pattern: '^\d{6}$'
          example: '654321'
      required: [code]

    # ── Profile / Preferences ─────────────────────────────────────────────────
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: 이메일 회원가입
      description: |
        이메일+비밀번호로 계정을 생성한다. 필수 약관 동의 및 연령 동의 필수.
        성공 시 액세스·리프레시 토큰을 반환한다(이메일 미인증 상태로 발급).
        계정 생성 트랜잭션 커밋 후 이메일 인증 코드 발송을 시도한다.
        이메일 발송 실패 시 계정은 유지되고, verificationEmailSent=false로 반환된다(항상 201).
        사용자는 이후 /auth/email-verification/request로 재발송할 수 있다.
        503/orphan-코드/한도 정책은 requestCode(재발송) 레벨에서 유지되며, signup은 계정을 항상 보존한다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 계정 생성 완료
          content:
            application/json:
              schema:
                type: object
                required: [account, tokens, verificationEmailSent]
                properties:
                  account:
                    $ref: '#/components/schemas/AccountSummary'
                  tokens:
                    $ref: '#/components/schemas/TokenPair'
                  verificationEmailSent:
                    type: boolean
                    description: 이메일 인증 코드 발송 성공 여부. 이메일 서비스 장애 시 false(계정은 생성됨). 재발송은 /auth/email-verification/request 이용.
        '409':
          description: 이메일 중복
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 유효성 검사 실패 (비밀번호 정책·약관 미동의·연령 미달)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 이메일 로그인
      description: |
        이메일+비밀번호로 인증 후 토큰 쌍을 반환한다.
        연속 5회 실패 시 계정 잠금 (30분 자동 해제, FR-019).
        잘못된 자격증명·계정 미존재·계정 잠금은 동일 401로 반환해 계정 열거를 방지함 (FR-026).
        소셜 전용 계정(비밀번호 없음)에는 422(SOCIAL_ONLY_ACCOUNT)로 소셜 로그인 안내 반환 — FR-026 수용 예외.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  account:
                    $ref: '#/components/schemas/AccountSummary'
                  tokens:
                    $ref: '#/components/schemas/TokenPair'
        '401':
          description: 인증 실패 (잘못된 자격증명·계정 미존재·계정 잠금 — 모두 동일 포맷으로 계정 열거 방지)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 계정 정지 (SUSPENDED 상태)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 소셜 전용 계정 — 비밀번호 미보유, 소셜 로그인 이용 안내 (code=SOCIAL_ONLY_ACCOUNT, FR-026 예외)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 액세스 토큰 갱신 (Refresh Token Rotation)
      description: |
        유효한 리프레시 토큰으로 새 토큰 쌍을 발급한다.
        기존 리프레시 토큰은 즉시 무효화된다(Rotation).
        이미 무효화된 토큰 재사용 감지 시 해당 family의 모든 세션을 무효화한다 (FR-020).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 토큰 갱신 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: 리프레시 토큰 만료 또는 무효화됨 (재로그인 필요)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃 (현재 세션 무효화)
      description: 전달한 리프레시 토큰에 해당하는 세션만 무효화한다. 다른 기기 세션은 유지.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: 로그아웃 성공 (본문 없음)
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/email-verification/request:
    post:
      tags: [Auth - Email Verification]
      summary: 이메일 인증 코드 재발송
      description: |
        가입 이메일로 인증 코드를 재발송한다. 기존 미사용 코드는 즉시 무효화.
        시간당 5회 초과 시 429. 이미 인증된 계정에서 호출 시 422.
        이메일 서비스 장애 시 503, 한도 미차감.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequestBody'
      responses:
        '202':
          description: 코드 발송 처리 완료
        '422':
          description: 이미 인증 완료된 계정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 재전송 횟수 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 이메일 서비스 장애
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/email-verification/verify:
    post:
      tags: [Auth - Email Verification]
      summary: 이메일 인증 코드 검증
      description: |
        이메일로 받은 6자리 코드를 검증하고 계정을 '인증' 상태로 전환한다.
        코드 만료 시 410, 오입력 5회 초과 시 코드 무효화.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationVerifyRequest'
      responses:
        '200':
          description: 이메일 인증 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  emailVerified:
                    type: boolean
                    example: true
        '401':
          description: 잘못된 코드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: 코드 만료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # ME — Authenticated + Email Verified
  # ═══════════════════════════════════════════════════════════════════════════
...