# src/main/java/com/newscurator/dto/request/LoginRequest.java
package com.newscurator.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
    @NotBlank @Email String email,
    @NotBlank String password
) {}